Separated server and routing
